+++ src/PhpSpreadsheet/Calculation/LookupRef/Helpers.php
@@ -61,8 +61,7 @@ class Helpers
     {
         $sheetName = '';
         if (str_contains($cellAddress, '!')) {
-            [$sheetName, $cellAddress] = Worksheet::extractSheetTitle($cellAddress, true);
-            $sheetName = trim($sheetName, "'");
+            [$sheetName, $cellAddress] = Worksheet::extractSheetTitle($cellAddress, true, true);
         }
 
         $worksheet = ($sheetName !== '')