+++ tests/PhpSpreadsheetTests/Calculation/CalculationTest.php
@@ -343,7 +343,7 @@ class CalculationTest extends TestCase
         // this used to raise a parser error, we keep it even though we don't
         // test the output
         $calculation->parseFormula($formula);
-        self::assertTrue(true);
+        self::assertSame(1, $calculation->cyclicFormulaCount);
     }
 
     public function testBranchPruningFormulaParsingInequalitiesConditionsCase(): void