+++ tests/PhpSpreadsheetTests/Reader/Ods/OdsInfoTest.php
@@ -70,6 +70,7 @@ class OdsInfoTest extends TestCase
                 'lastColumnIndex' => 2,
                 'totalRows' => 12,
                 'totalColumns' => 3,
+                'sheetState' => 'visible',
             ],
             [
                 'worksheetName' => 'Second Sheet',
@@ -77,6 +78,7 @@ class OdsInfoTest extends TestCase
                 'lastColumnIndex' => 0,
                 'totalRows' => 2,
                 'totalColumns' => 1,
+                'sheetState' => 'visible',
             ],
         ], $wsinfo);
     }