+++ tests/PhpSpreadsheetTests/Worksheet/RowCellIteratorTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace PhpOffice\PhpSpreadsheetTests\Worksheet;
 
-use PhpOffice\PhpSpreadsheet\Cell\Cell;
 use PhpOffice\PhpSpreadsheet\Exception as Except;
 use PhpOffice\PhpSpreadsheet\Spreadsheet;
 use PhpOffice\PhpSpreadsheet\Worksheet\RowCellIterator;
@@ -41,10 +40,8 @@ class RowCellIteratorTest extends TestCase
 
         $values = [];
         foreach ($iterator as $key => $RowCell) {
-            self::assertNotNull($RowCell);
             $values[] = $RowCell->getValue();
             self::assertEquals($RowCellIndexResult++, $key);
-            self::assertInstanceOf(Cell::class, $RowCell);
         }
         self::assertSame(self::CELL_VALUES[0], $values);
         $spreadsheet->disconnectWorksheets();
@@ -60,10 +57,8 @@ class RowCellIteratorTest extends TestCase
 
         $values = [];
         foreach ($iterator as $key => $RowCell) {
-            self::assertNotNull($RowCell);
             $values[] = $RowCell->getValue();
             self::assertEquals($RowCellIndexResult++, $key);
-            self::assertInstanceOf(Cell::class, $RowCell);
         }
         self::assertSame([220, 230, 240], $values);
         $spreadsheet->disconnectWorksheets();